video/VLM/requirements.txt
torch>=2.0.0
Pillow>=9.0.0
numpy>=1.24.0
opencv-python>=4.8.0
tqdm>=4.0.0
networkx>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
bitsandbytes>=0.39.0 